# Startup flutter faster (faster wrapper script on Windows)